+++ CHANGELOG.md
@@ -4,6 +4,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html). See [MAINTAINERS.md](./MAINTAINERS.md)
 for instructions to keep up to date.
 
+## UNRELEASED
+
+### FIXED
+
+* Substreams running without a specific tier2 (with default `substreams-client-endpoint=<empty>`)  will now expose tier2 service `sf.substreams.internal.v2.Substreams` so it can be used internally.
+
+### CHANGED
+
+* flag `substreams-client-endpoint` now defaults to empty string, which will make a 'tier1' service act also as a 'tier2', serving its own requests. Make sure that you don't expose `sf.substreams.internal.v2.Substreams` to the public, via an HTTP-aware firewall.
+* flag `substreams-partial-mode-enabled` renamed to `substreams-tier2`
+
 ## v1.4.0
 
 ### Substreams RPC protocol V2